+++ modules/validators/builtin/include/mustBeNonSparseBuiltin.hpp
@@ -26,16 +26,13 @@
 #pragma once
 //=============================================================================
 #include "ArrayOf.hpp"
-#include "Evaluator.hpp"
-#include "nlsTypes_builtin_exports.h"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-namespace TypeGateway {
+namespace ValidatorsGateway {
     //=============================================================================
-    NLSTYPES_BUILTIN_IMPEXP
     ArrayOfVector
-    issparseBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+    mustBeNonSparseBuiltin(int nLhs, const ArrayOfVector& argIn);
     //=============================================================================
 }
 //=============================================================================