+++ modules/validators/src/include/Validators.hpp
@@ -87,5 +87,8 @@ mustBeNonNan(const ArrayOfVector& args, int argPosition);
 NLSVALIDATORS_IMPEXP void
 mustBeNonZero(const ArrayOfVector& args, int argPosition);
 //=============================================================================
+NLSVALIDATORS_IMPEXP void
+mustBeNonSparse(const ArrayOfVector& args, int argPosition);
+//=============================================================================
 }
 //=============================================================================